requirements.txt
mplfinance==0.12.10b0
numpy==2.2.1
pandas==2.2.3
python-decouple==3.8
SQLAlchemy==2.0.37
websocket-client==1.8.0
ta~=0.11.0
python-binance==1.0.27
python-decouple==3.8
fastapi==0.115.8
uvicorn==0.34.0
requests==2.32.3
